+++ src/main/java/com/yohanii/lostandfound/domain/image/Image.java
@@ -1,7 +1,7 @@
 package com.yohanii.lostandfound.domain.image;
 
 import com.yohanii.lostandfound.domain.item.Item;
-import com.yohanii.lostandfound.domain.user.User;
+import com.yohanii.lostandfound.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -21,8 +21,8 @@ public class Image {
     private Long id;
 
     @OneToOne(fetch = LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
+    @JoinColumn(name = "member_id")
+    private Member member;
 
     @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "item_id")
@@ -36,8 +36,8 @@ public class Image {
     private String storeFileName;
 
     @Builder
-    public Image(User user, Item item, ImageType type, String uploadFileName, String storeFileName) {
-        this.user = user;
+    public Image(Member member, Item item, ImageType type, String uploadFileName, String storeFileName) {
+        this.member = member;
         this.item = item;
         this.type = type;
         this.uploadFileName = uploadFileName;