+++ src/main/java/com/yohanii/lostandfound/domain/image/ImageType.java
@@ -1,5 +1,5 @@
 package com.yohanii.lostandfound.domain.image;
 
 public enum ImageType {
-    GENERAL, USER, ITEM
+    GENERAL, MEMBER, ITEM
 }