+++ src/main/java/com/yohanii/lostandfound/domain/member/Member.java
@@ -1,4 +1,4 @@
-package com.yohanii.lostandfound.domain.user;
+package com.yohanii.lostandfound.domain.member;
 
 import com.yohanii.lostandfound.domain.image.Image;
 import com.yohanii.lostandfound.dto.profile.ProfileEditRequestDto;
@@ -10,17 +10,16 @@ import java.time.LocalDateTime;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
 @Entity
-@Table(name = "users")
-public class User {
+public class Member {
 
     @Id @GeneratedValue
-    @Column(name = "user_id")
+    @Column(name = "member_id")
     private Long id;
 
-    @Column(name = "user_name")
+    @Column(name = "member_name")
     private String name;
 
-    @OneToOne(mappedBy = "user", fetch = FetchType.LAZY)
+    @OneToOne(mappedBy = "member", fetch = FetchType.LAZY)
     private Image profileImage;
 
     private String loginId;
@@ -30,7 +29,7 @@ public class User {
     private LocalDateTime updatedTime;
 
     @Builder
-    public User(String name, Image profileImage, String loginId, String password, String nickName, LocalDateTime createdTime, LocalDateTime updatedTime) {
+    public Member(String name, Image profileImage, String loginId, String password, String nickName, LocalDateTime createdTime, LocalDateTime updatedTime) {
         this.name = name;
         this.profileImage = profileImage;
         this.loginId = loginId;
@@ -40,7 +39,7 @@ public class User {
         this.updatedTime = updatedTime;
     }
 
-    public void updateUser(ProfileEditRequestDto dto) {
+    public void updateMember(ProfileEditRequestDto dto) {
         this.name = dto.getName();
         this.nickName = dto.getNickName();
         this.updatedTime = LocalDateTime.now();