+++ src/main/java/com/yohanii/lostandfound/domain/post/Post.java
@@ -1,7 +1,7 @@
 package com.yohanii.lostandfound.domain.post;
 
 import com.yohanii.lostandfound.domain.item.Item;
-import com.yohanii.lostandfound.domain.user.User;
+import com.yohanii.lostandfound.domain.member.Member;
 import com.yohanii.lostandfound.dto.post.PostEditRequestDto;
 import jakarta.persistence.*;
 import lombok.*;
@@ -21,8 +21,8 @@ public class Post {
     private Long id;
 
     @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
+    @JoinColumn(name = "member_id")
+    private Member member;
 
     @OneToOne(mappedBy = "post", fetch = FetchType.LAZY)
     private Item item;
@@ -41,8 +41,8 @@ public class Post {
     private LocalDateTime updatedTime;
 
     @Builder
-    public Post(User user, Item item, String title, String content, PostType type, LocalDateTime createdTime, LocalDateTime updatedTime) {
-        this.user = user;
+    public Post(Member member, Item item, String title, String content, PostType type, LocalDateTime createdTime, LocalDateTime updatedTime) {
+        this.member = member;
         this.item = item;
         this.title = title;
         this.content = content;