+++ src/main/java/com/yohanii/lostandfound/service/login/LoginService.java
@@ -1,7 +1,7 @@
 package com.yohanii.lostandfound.service.login;
 
-import com.yohanii.lostandfound.domain.user.User;
-import com.yohanii.lostandfound.domain.user.UserRepository;
+import com.yohanii.lostandfound.domain.member.Member;
+import com.yohanii.lostandfound.domain.member.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,12 +13,12 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class LoginService {
 
-    private final UserRepository userRepository;
+    private final MemberRepository memberRepository;
 
-    public Optional<User> login(String loginId, String password) {
-        User findUser = userRepository.findByLoginId(loginId)
-                .filter(user -> user.getPassword().equals(password))
+    public Optional<Member> login(String loginId, String password) {
+        Member findMember = memberRepository.findByLoginId(loginId)
+                .filter(member -> member.getPassword().equals(password))
                 .orElse(null);
-        return Optional.ofNullable(findUser);
+        return Optional.ofNullable(findMember);
     }
 }