+++ src/main/java/com/yohanii/lostandfound/web/SessionConst.java
@@ -1,5 +1,5 @@
 package com.yohanii.lostandfound.web;
 
 public interface SessionConst {
-    String LOGIN_USER = "loginUser";
+    String LOGIN_MEMBER = "loginMember";
 }