+++ src/test/java/com/yohanii/lostandfound/domain/post/PostRepositoryTest.java
@@ -1,16 +1,13 @@
 package com.yohanii.lostandfound.domain.post;
 
-import com.yohanii.lostandfound.domain.user.User;
-import com.yohanii.lostandfound.domain.user.UserRepository;
+import com.yohanii.lostandfound.domain.member.Member;
+import com.yohanii.lostandfound.domain.member.MemberRepository;
 import com.yohanii.lostandfound.dto.post.PostSearchRequestDto;
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.jdbc.EmbeddedDatabaseConnection;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -25,13 +22,13 @@ class PostRepositoryTest {
     @Autowired
     PostRepository postRepository;
     @Autowired
-    UserRepository userRepository;
+    MemberRepository memberRepository;
 
     @Test
     void save() {
-        User user = User.builder().build();
+        Member member = Member.builder().build();
         Post post = Post.builder()
-                .user(user)
+                .member(member)
                 .title("hello")
                 .content("hello123")
                 .type(PostType.LOST)
@@ -40,7 +37,7 @@ class PostRepositoryTest {
         Long savedId = postRepository.save(post);
 
         Post savedPost = postRepository.findById(savedId);
-        assertThat(savedPost.getUser()).isEqualTo(user);
+        assertThat(savedPost.getMember()).isEqualTo(member);
         assertThat(savedPost.getTitle()).isEqualTo("hello");
         assertThat(savedPost.getContent()).isEqualTo("hello123");
         assertThat(savedPost.getType()).isEqualTo(PostType.LOST);
@@ -161,15 +158,15 @@ class PostRepositoryTest {
     }
 
     @Test
-    void findAllByUserId() {
-        User user = User.builder().build();
-        userRepository.save(user);
+    void findAllByMemberId() {
+        Member member = Member.builder().build();
+        memberRepository.save(member);
 
         Post post1 = Post.builder()
-                .user(user)
+                .member(member)
                 .build();
         Post post2 = Post.builder()
-                .user(user)
+                .member(member)
                 .build();
         Post post3 = Post.builder().build();
 
@@ -177,7 +174,7 @@ class PostRepositoryTest {
         postRepository.save(post2);
         postRepository.save(post3);
 
-        List<Post> findPosts = postRepository.findAll(user.getId());
+        List<Post> findPosts = postRepository.findAll(member.getId());
 
         assertThat(findPosts.size()).isEqualTo(2);
         assertThat(findPosts).contains(post1, post2);